Create Zoo Keeper Final Project
There's this zoo. It has been installed with 6 cameras (0 - 5). This program asks the user which habitat do they want to view (there's a difference in wording because hyperskill needs the wording (strings) to be exactly what they say, down to every letter, or their online learning platform won't accept the code). So, anyway, the user enters a camera from 0 to 5 and this will show the corresponding habitat e.g. lion, rabbit, swan, bat. It will show you the animal. It will then loop back to the original question "Which habitat # do you need?" so the user can change cameras to view a different habitat. This repeats until the user types "exit".