Create particles: a continuously running database with an exit option
Two important characteristics of the elementary particles are the spin and the electric charge. Here are some of the elementary particles:

Particle	Class	Spin	Electric charge
Strange	Quark	1/2	-1/3
Charm	Quark	1/2	2/3
Electron	Lepton	1/2	-1
Neutrino	Lepton	1/2	0
Photon	Boson	1	0

I've written a program that retrieves the Particle and Class based on the user inputting the Spin and Charge. The program also lets the user know if their is no record of the particle in the database (the database is the table above). Lastly, at any time the user can enter the words exit and the program will stop. Otherwise the program will continue running.